docs: unlink GoDoc examples from readme
I want the examples in the readme to stand reflect what a user might
actually write in their test files. Due to GoDoc's requirements for an
example file, this can be difficult to do with test-specific packages
like sqlitestdb.

This changeset unlinks the GoDoc examples from the readme examples,
allowing the readme to reflect what a user will actually write, while
still having some example in the GoDocs.