fix(libsql): workaround bugs in go driver
The libSQL driver for Go has two bugs which affect how sqlitestdb
creates template databases, which was masked by it sharing a template
database with the SQLite drivers.

The first is that statements ran with `ExecContext` cannot return a row,
else a non-nil error is returned. When executing a PRAGMA statement
libSQL returns the set value as a row, trigging this bug. This differs
from all the other database/sql drivers where a nil error is returned in
this case.

The second bug is triggered also in `ExecContext` where multiple
statements are provided. The underlying `sqlite3_exec` C function
accepts multiple statements, and other drivers support this. However,
the libSQL driver only executes the first statement, all others are
silently ignored.

In this changeset the first bug is addressed by changing sqlitestdb to
use `QueryRowContext` and checking the error from `Row.Scan`. The second
bug is addressed in the libsql test package by separating each component
of the string migration.