chore: use cancelable context
In Go 1.24 the testing package will include a function to return a
context that is canceled at the end of the test. This is similar to, but
not exactly like, using WithCancel and calling `defer cancel()`.

As sqlitestdb and the tests should be handling the lifecycles of
resources correctly, this is a no-op. By opting into the behavior early,
I can ensure sqlitestdb continues to have no issues.